refactor(kody-rules): share splitRulePathGlobs between matcher and IDE-dir rejection (Kody review)
pathMatchesIdeRuleDir split comma-joined globs naively, diverging from
the brace/character-class-aware split used by the review matchers.
Extract splitRulePathGlobs to file-patterns.ts as the single source of
truth and use it at both sites.
